requirements.txt
pandas
numpy
matplotlib
seaborn
plotly
scipy
psycopg2-binary
streamlit
sqlalchemy
tabulate
python-dotenv
google-adk
requests
